Rename Maven group ID to 'io.micronaut.neo4j'
