+++ .github/workflows/dependency-update.yml
@@ -1,0 +1,37 @@
+# WARNING: Do not edit this file directly. Instead, go to:
+#
+# https://github.com/micronaut-projects/micronaut-project-template/tree/master/.github/workflows
+#
+# and edit them there. Note that it will be sync'ed to all the Micronaut repos
+name: Update Dependencies
+on:
+  schedule:
+    - cron: '0 6 * * *'
+jobs:
+  dependency-updates:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/cache@v1
+        with:
+          path: ~/.gradle/caches
+          key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle') }}
+          restore-keys: |
+            ${{ runner.os }}-gradle-
+      - name: Set up JDK
+        uses: actions/setup-java@v1
+        with:
+          java-version: 1.8
+      - name: Export Gradle Properties
+        uses: micronaut-projects/github-actions/export-gradle-properties@master
+      - name: Check Dependencies
+        run: ./gradlew useLatestVersions
+      - name: Create Pull Request
+        uses: peter-evans/create-pull-request@v2
+        with:
+          token: ${{ secrets.GH_TOKEN }}
+          commit-message: Update dependencies
+          title: 'Dependency upgrades'
+          body: Upgrades dependencies to their latest versions
+          labels: "type: dependency-upgrade"
+          base: ${{ env.githubBranch }}
