+++ .github/workflows/gradle.yml
@@ -1,42 +1,52 @@
+# WARNING: Do not edit this file directly. Instead, go to:
+#
+# https://github.com/micronaut-projects/micronaut-project-template/tree/master/.github/workflows
+#
+# and edit them there. Note that it will be sync'ed to all the Micronaut repos
 name: Java CI
 on:
   push:
     branches:
-    - master
+      - master
+      - '[1-9]+.[0-9]+.x'
   pull_request:
     branches:
-    - master   
+      - master
+      - '[1-9]+.[0-9]+.x'
 jobs:
   build:
+    if: github.repository != 'micronaut-projects/micronaut-project-template'
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        java: ['1.8', '11.0.4']
+        java: ['8', '11', '14']
     steps:
-    - uses: actions/checkout@v1
-    - uses: actions/cache@v1
-      with:
-        path: ~/.gradle/caches
-        key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle') }}
-        restore-keys: |
-          ${{ runner.os }}-gradle-
-    - name: Set up JDK
-      uses: actions/setup-java@v1
-      with:
-        java-version: ${{ matrix.java }}   
-    - name: Build with Gradle
-      run: ./gradlew check --continue
-    - name: Publish to JFrog OSS
-      if: success() && github.event_name == 'push' && matrix.java == '1.8'
-      env:
-         BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
-         BINTRAY_KEY: ${{ secrets.BINTRAY_KEY }}
-      run: ./gradlew publish docs
-    - name: Publish to Github Pages
-      if: success() && github.event_name == 'push' && matrix.java == '1.8'
-      uses: micronaut-projects/github-pages-deploy-action@master
-      env:
-        GH_TOKEN: ${{ secrets.GH_TOKEN }}
-        BASE_BRANCH: master 
-        BRANCH: gh-pages 
-        FOLDER: build/docs
+      - uses: actions/checkout@v2
+      - uses: actions/cache@v1
+        with:
+          path: ~/.gradle/caches
+          key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle') }}
+          restore-keys: |
+            ${{ runner.os }}-gradle-
+      - name: Set up JDK
+        uses: actions/setup-java@v1
+        with:
+          java-version: ${{ matrix.java }}
+      - name: Build with Gradle
+        run: ./gradlew dependencyUpdates check --parallel --continue
+        env:
+          TESTCONTAINERS_RYUK_DISABLED: true
+      - name: Publish to JFrog OSS
+        if: success() && github.event_name == 'push' && matrix.java == '8'
+        env:
+          BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
+          BINTRAY_KEY: ${{ secrets.BINTRAY_KEY }}
+        run: ./gradlew publish docs
+      - name: Publish to Github Pages
+        if: success() && github.event_name == 'push' && matrix.java == '8'
+        uses: micronaut-projects/github-pages-deploy-action@master
+        env:
+          GH_TOKEN: ${{ secrets.GH_TOKEN }}
+          BASE_BRANCH: ${{ env.githubBranch }}
+          BRANCH: gh-pages
+          FOLDER: build/docs
