+++ .github/workflows/label-sync.yml
@@ -1,0 +1,21 @@
+# WARNING: Do not edit this file directly. Instead, go to:
+#
+# https://github.com/micronaut-projects/micronaut-project-template/tree/master/.github/workflows
+#
+# and edit them there. Note that it will be sync'ed to all the Micronaut repos
+name: Sync labels
+on:
+  issues:
+    types: [opened, edited, deleted, labeled, closed]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Download labels.yml
+        run: curl -O https://raw.githubusercontent.com/micronaut-projects/micronaut-build/master/labels.yml
+      - uses: micnncim/action-label-syncer@v1
+        env:
+          GITHUB_TOKEN: ${{ secrets.GH_TOKEN }}
+        with:
+          manifest: labels.yml
