+++ .github/workflows/release-notes.yml
@@ -1,0 +1,50 @@
+# WARNING: Do not edit this file directly. Instead, go to:
+#
+# https://github.com/micronaut-projects/micronaut-project-template/tree/master/.github/workflows
+#
+# and edit them there. Note that it will be sync'ed to all the Micronaut repos
+name: Release Notes
+on:
+  pull_request:
+    types: [closed, labeled]
+    branches:
+      - master
+      - '[1-9]+.[0-9]+.x'
+  issues:
+    types: [closed,reopened, labeled]
+  push:
+    branches:
+      - master
+      - '[1-9]+.[0-9]+.x'
+    paths:
+      - ".github/workflows/release-notes.yml"
+jobs:
+  release_notes:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Export Gradle Properties
+        uses: micronaut-projects/github-actions/export-gradle-properties@master
+      - uses: micronaut-projects/github-actions/release-notes@master
+        id: release_notes
+        with:
+          branch: ${{ env.githubBranch }}
+      - uses: ncipollo/release-action@v1
+        if: success()
+        with:
+          allowUpdates: true
+          commit: ${{ env.githubBranch }}
+          draft: true
+          name: ${{ env.title }} ${{ steps.release_notes.outputs.next_version }}
+          tag: v${{ steps.release_notes.outputs.next_version }}
+          bodyFile: CHANGELOG.md
+          token: ${{ secrets.GITHUB_TOKEN }}
+
+  release_drafter:
+    runs-on: ubuntu-latest
+    needs: release_notes
+    steps:
+      - uses: release-drafter/release-drafter@v5
+        if: failure()
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
