+++ .github/workflows/release.yml
@@ -1,37 +1,64 @@
+# WARNING: Do not edit this file directly. Instead, go to:
+#
+# https://github.com/micronaut-projects/micronaut-project-template/tree/master/.github/workflows
+#
+# and edit them there. Note that it will be sync'ed to all the Micronaut repos
 name: Release
 on:
-  push:
-    tags:
-    - v*
+  release:
+    types: [published]
 jobs:
-  build:
+  release:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v1
-    - name: Set up JDK
-      uses: actions/setup-java@v1
-      with:
-        java-version: 1.8
-        ref: ${{ github.head_ref }}
-    - name: Publish to Bintray
-      env:
-         BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
-         BINTRAY_KEY: ${{ secrets.BINTRAY_KEY }}
-      run: ./gradlew bintrayUpload docs
-    - id: release_version
-      run: echo ::set-output name=release_version::${GITHUB_REF:11}
-    - name: Publish to Github Pages
-      if: success()
-      uses: micronaut-projects/github-pages-deploy-action@master
-      env:
-        GH_TOKEN: ${{ secrets.GH_TOKEN }}
-        BASE_BRANCH: master 
-        BRANCH: gh-pages 
-        FOLDER: build/docs
-        VERSION: ${{ steps.release_version.outputs.release_version }}
-    - name: Synchronize Maven Central
-      if: success()
-      env:
-         BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
-         BINTRAY_KEY: ${{ secrets.BINTRAY_KEY }}
-      run: ./gradlew sWMC
+      - name: Checkout repository
+        uses: actions/checkout@v2
+      - name: Set up JDK
+        uses: actions/setup-java@v1
+        with:
+          java-version: 1.8
+      - name: Set the current release version
+        id: release_version
+        run: echo ::set-output name=release_version::${GITHUB_REF:11}
+      - name: Run pre-release
+        uses: micronaut-projects/github-actions/pre-release@master
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+      - name: Publish to Bintray
+        env:
+          BINTRAY_USER: ${{ secrets.BINTRAY_USER }}
+          BINTRAY_KEY: ${{ secrets.BINTRAY_KEY }}
+        run: ./gradlew bintrayUpload docs
+      - name: Export Gradle Properties
+        uses: micronaut-projects/github-actions/export-gradle-properties@master
+      - name: Publish to Github Pages
+        if: success()
+        uses: micronaut-projects/github-pages-deploy-action@master
+        env:
+          BETA: ${{ contains(steps.release_version.outputs.release_version, 'M') || contains(steps.release_version.outputs.release_version, 'RC') }}
+          GH_TOKEN: ${{ secrets.GH_TOKEN }}
+          BASE_BRANCH: ${{ env.githubBranch }}
+          BRANCH: gh-pages
+          FOLDER: build/docs
+          VERSION: ${{ steps.release_version.outputs.release_version }}
+      - name: Checkout micronaut-core
+        uses: actions/checkout@v2
+        with:
+          token: ${{ secrets.GH_TOKEN }}
+          repository: micronaut-projects/micronaut-core
+          ref: ${{ env.githubCoreBranch }}
+          path: micronaut-core # Must be micronaut-core
+        continue-on-error: true
+      - name: Update BOM
+        uses: micronaut-projects/github-actions/update-bom@master
+        with:
+          token: ${{ secrets.GH_TOKEN }}
+        continue-on-error: true
+      - name: Run post-release
+        if: success()
+        uses: micronaut-projects/github-actions/post-release@master
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+   
+
+
