+++ gradle.properties
@@ -1,17 +1,27 @@
-projectVersion=2.0.1.BUILD-SNAPSHOT
-kafkaVersion=2.1.0
-micronautDocsVersion=1.0.22
-micronautBuildVersion=1.0.0
-micrometerVersion=1.3.1
+projectVersion=3.0.0.BUILD-SNAPSHOT
+neo4jDriverVersion=4.0.1
+neo4jVersion=3.5.16
+micronautDocsVersion=1.0.23
+micronautBuildVersion=1.1.5
+micrometerVersion=1.3.5
 validationVersion=2.0.1.Final
-micronautVersion=1.2.8
-micronautTestVersion=1.1.3
-groovyVersion=2.5.7
+micronautVersion=1.3.5
+micronautTestVersion=1.1.5
+groovyVersion=2.5.11
 title=Micronaut Neo4j
 projectDesc=Integration between Micronaut and Neo4j
 projectUrl=http://micronaut.io
 githubSlug=micronaut-projects/micronaut-neo4j
+
+# This project's branch
 githubBranch=master
+
+# Micronaut core branch for BOM pull requests
+githubCoreBranch=master
+
+bomProperty=micronautNeo4jVersion
+# If needed, set additional properties
+bomProperties=neo4jDriverVersion,neo4jVersion
 spockVersion=1.2-groovy-2.5
 developers=Graeme Rocher
 # Needed for documentation generation. Needs to be Grails 3.2.x not 3.3.x
