+++ neo4j-bolt/src/main/java/io/micronaut/neo4j/bolt/Neo4jBoltSettings.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2017-2018 original authors
+ * Copyright 2017-2020 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,11 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.micronaut.neo4j.bolt;
 
 /**
- * Neo4j condition.
- *
  * @author graemerocher
  * @since 1.0
  */
-package io.micronaut.configuration.neo4j.bolt.condition;
+public interface Neo4jBoltSettings {
+    /**
+     * The prefix to use for all Neo4j settings.
+     */
+    String PREFIX = "neo4j";
+    /**
+     * Default Neo4j URI.
+     */
+    String DEFAULT_URI = "bolt://localhost:7687";
+    /**
+     * The default embedded data location.
+     */
+    String DEFAULT_LOCATION = "data/neo4j";
+}
