+++ neo4j-bolt/src/main/java/io/micronaut/neo4j/bolt/condition/RequiresNeo4j.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2017-2018 original authors
+ * Copyright 2017-2020 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.micronaut.neo4j.bolt.condition;
 
-package io.micronaut.configuration.neo4j.bolt.condition;
-
-import io.micronaut.configuration.neo4j.bolt.Neo4jBoltSettings;
+import io.micronaut.neo4j.bolt.Neo4jBoltSettings;
 import io.micronaut.context.annotation.Requires;
 import org.neo4j.driver.Config;
