+++ neo4j-bolt/src/test/groovy/io/micronaut/neo4j/bolt/Neo4jBoltConfigurationSpec.groovy
@@ -14,31 +14,28 @@
  * limitations under the License.
  */
 
-package io.micronaut.configuration.neo4j.bolt
+package io.micronaut.neo4j.bolt
 
-import org.neo4j.driver.Driver
 import io.micronaut.context.ApplicationContext
-import io.micronaut.core.io.socket.SocketUtils
 import spock.lang.Specification
 
 /**
  * @author graemerocher
  * @since 1.0
  */
-class Neo4jEmbeddedServerSpec extends Specification{
+class Neo4jBoltConfigurationSpec extends Specification {
 
-    void "test neo4j embedded"() {
-        given:
+    void "test neo4j configuration"() {
+        when:
         ApplicationContext applicationContext = ApplicationContext.run(
-                'neo4j.uri':"bolt://localhost:${SocketUtils.findAvailableTcpPort()}",
-                'neo4j.embedded.ephemeral':true
+                'neo4j.uri':'bolt://someserver:7687',
+                'neo4j.embedded.enabled':false
         )
 
-        when:
-        Driver driver = applicationContext.getBean(Driver)
-
         then:
-        driver.session().run('MATCH (n) RETURN n').size() == 0
+        applicationContext.containsBean(Neo4jBoltConfiguration)
+        applicationContext.getBean(Neo4jBoltConfiguration).uris.size() == 1
+        applicationContext.getBean(Neo4jBoltConfiguration).uris[0] == URI.create('bolt://someserver:7687')
 
         cleanup:
         applicationContext?.stop()
