+++ src/main/docs/guide/config.adoc
@@ -1,4 +1,4 @@
-The configuration can be further customized with all the options available using the api:configuration.neo4j.bolt.Neo4jBoltConfiguration[] class.
+The configuration can be further customized with all the options available using the api:neo4j.bolt.Neo4jBoltConfiguration[] class.
 
 The `Neo4jBoltConfiguration` also exposes all the options for the `org.neo4j.driver.v1.Config.ConfigBuilder` class.
