+++ src/main/docs/guide/setup.adoc
@@ -9,7 +9,7 @@ $ mn create-app my-app --features neo4j-bolt
 
 To configure the Neo4j Bolt driver you should first add the `neo4j-bolt` module to your classpath:
 
-dependency::micronaut-neo4j-bolt[groupId="io.micronaut.configuration", version="{version}"]
+dependency::micronaut-neo4j-bolt[groupId="io.micronaut.neo4j", version="{version}"]
 
 NOTE: The 1.x line of `micronaut-neo4j-bolt` supports Neo4j Java Driver 1.x. The latest version of the Neo4j Java Driver is the 4.x line which changes the package name (amongst other breaking API changes) therefore in order to follow semantic versiooning the 2.x line of `micronaut-neo4j-bolt` is dedicated to this version.
 
@@ -29,6 +29,6 @@ Once you have the above configuration in place you can inject an instance of the
 
 === Neo4j Health Checks
 
-When the `micronaut-neo4j-bolt` module is activated a api:configuration.neo4j.bolt.health.Neo4jHealthIndicator[] is activated resulting in the `/health` endpoint and https://docs.micronaut.io/latest/api/io/micronaut/health/CurrentHealthStatus.html[CurrentHealthStatus] interface resolving the health of the Neo4j connection.
+When the `micronaut-neo4j-bolt` module is activated a api:neo4j.bolt.health.Neo4jHealthIndicator[] is activated resulting in the `/health` endpoint and https://docs.micronaut.io/latest/api/io/micronaut/health/CurrentHealthStatus.html[CurrentHealthStatus] interface resolving the health of the Neo4j connection.
 
 See the section on the https://docs.micronaut.io/latest/guide/index.html#healthEndpoint[Health Endpoint] for more information.
