+++ src/main/docs/guide/testing.adoc
@@ -4,7 +4,7 @@ dependency:neo4j-harness[groupId="org.neo4j.test", scope="test"]
 
 If the Neo4j server is not already running on the configured port an embedded version will be started.
 
-You can configure the options for the embedded Neo4j server using the `neo4j.embedded` settings exposed by api:configuration.neo4j.bolt.Neo4jBoltConfiguration[].
+You can configure the options for the embedded Neo4j server using the `neo4j.embedded` settings exposed by api:neo4j.bolt.Neo4jBoltConfiguration[].
 
 One useful option is `ephemeral` which ensures the data is cleaned up between test runs. For example in `application-test.yml`:
